refactor(Balance): Put new components in components directory
