+++ src/ducks/balance/components/AccountsList.styl
@@ -1,9 +1,3 @@
-.GroupPannelSummary__figure
-    padding-right 0 !important
-
-.GroupPannelSummary__figureCurrency
-    color inherit
-
 .AccountsList
     flex 1
     margin 0
