+++ src/ducks/balance/components/GroupPanel.styl
@@ -1,0 +1,5 @@
+.GroupPannelSummary__figure
+    padding-right 0 !important
+
+.GroupPannelSummary__figureCurrency
+    color inherit
